small change  on migrations

+++ db/migrate/20191106200853_change_roles_table.rb
@@ -1,37 +1,21 @@
 class ChangeRolesTable < ActiveRecord::Migration[5.1]
   def up
-    add_column :roles, :post, :integer, default: 0, null: false
-    add_column :roles, :chat, :integer, default: 0, null: false
-    add_column :roles, :event, :integer, default: 0, null: false
-    add_column :roles, :merchandise, :integer, default: 0, null: false
-    add_column :roles, :badge, :integer, default: 0, null: false
-    add_column :roles, :reward, :integer, default: 0, null: false
-    add_column :roles, :quest, :integer, default: 0, null: false
-    add_column :roles, :beacon, :integer, default: 0, null: false
-    add_column :roles, :reporting, :integer, default: 0, null: false
-    add_column :roles, :interest, :integer, default: 0, null: false
-    add_column :roles, :courseware, :integer, default: 0, null: false
-    add_column :roles, :trivia, :integer, default: 0, null: false
-    add_column :roles, :admin, :integer, default: 0, null: false
-    add_column :roles, :root, :integer, default: 0, null: false
-    remove_column :roles, :role_enum
+
+    [:post, :chat, :event, :merchandise, :badge,:reward, :quest,:beacon, :reporting,:interest, :courseware, :trivia, :admin, :root].each do |column|
+      unless column_exists? :roles, column
+        add_column :roles, column, :integer, default: 0, null: false
+      end
+    end
+
+    if column_exists? :roles, :role_enum
+      remove_column :roles, :role_enum
+    end
   end
 
   def down
-    remove_column :roles, :post
-    remove_column :roles, :chat
-    remove_column :roles, :event
-    remove_column :roles, :merchandise
-    remove_column :roles, :badge
-    remove_column :roles, :reward
-    remove_column :roles, :quest
-    remove_column :roles, :beacon
-    remove_column :roles, :reporting
-    remove_column :roles, :interest
-    remove_column :roles, :courseware
-    remove_column :roles, :trivia
-    remove_column :roles, :admin
-    remove_column :roles, :root
+    [:post, :chat, :event, :merchandise, :badge,:reward, :quest,:beacon, :reporting,:interest, :courseware, :trivia, :admin, :root].each do |column|
+      remove_column :roles, column
+    end
     add_column :roles, :role_enum, :integer, default: 0
 
   end
